+++ tests/unittests/test_health.py
@@ -1,7 +1,13 @@
-def test_health(app_client):
-    app_client.get('/api/health').raise_for_status()
+from pytest import mark
 
 
-def test_unhealthy(app_client, mock_engine):
+@mark.asyncio
+async def test_health(app_client):
+    (await app_client.get('/api/health')).raise_for_status()
+
+
+@mark.asyncio
+async def test_unhealthy(app_client, mock_engine):
     mock_engine.connection.execute.side_effect = Exception
-    assert app_client.get('/api/health').status_code == 500
+    app_client.application.health_monitor.status = await app_client.application.health_monitor._check()
+    assert (await app_client.get('/api/health')).status_code == 500